# Conflicts:
#	CHANGELOG.rst

+++ tests/integration/cases/float_special_values/Cpp_combined.cpp
@@ -3,13 +3,13 @@
 #include <vector>
 void check_() {
 auto my_data = std::vector<double>{
-    INFINITY,
-    -INFINITY,
-    NAN,
+    static_cast<double>(INFINITY),
+    -static_cast<double>(INFINITY),
+    static_cast<double>(NAN),
 };
 my_data = std::vector<double>{
-    INFINITY,
-    -INFINITY,
-    NAN,
+    static_cast<double>(INFINITY),
+    -static_cast<double>(INFINITY),
+    static_cast<double>(NAN),
 };
 }
